Automatically generate OAuth flow URI. Fixes #1.
I needed to change how the constructor set the options, in order to allow for
optional options (that seems a bit redundant). I extracted the setting of the
options into a method, and check required options, just to clean up the
constructor. Also, out of OCD, I moved the redirectUri variable so that the
application id and secret were next to each other, and this alphabetizes things
nicely :)